# docs/v2/contracts/interpretaai-v2.openapi.yaml
openapi: 3.1.0
info:
  title: InterpretaAI API v2
  version: 2.0.0-draft.1
  description: >-
    Contrato de projeto para autoria docente, publicação, entrega offline e evidências.
    Não representa uma API já implantada.
servers:
  - url: https://api.example.interpreta.ai/api/v2
    description: Endereço ilustrativo; nenhum ambiente é afirmado como ativo.
...
security:
  - adultOidc: []
paths:
...
  /device-management/pairing-codes:
    post:
      tags: [Device Management]
      operationId: createDevicePairingCode
      summary: Cria código efêmero e de uso único para uma turma autorizada
      parameters:
        - $ref: '#/components/parameters/SchoolContext'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePairingCodeRequest' }
      responses:
        '201':
          description: Código exibido uma única vez ao adulto
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PairingCode' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /device-pairings/redeem:
    post:
      tags: [Device Management]
      operationId: redeemDevicePairingCode
      summary: Troca o código efêmero por credencial própria do aparelho
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RedeemPairingRequest' }
      responses:
        '201':
          description: Credencial exibida uma única vez ao aparelho
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceCredential' }
        '400': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /device-management/devices/{deviceId}/revoke:
    post:
      tags: [Device Management]
      operationId: revokeDeviceCredential
      summary: Revoga imediatamente a credencial de um aparelho autorizado
      parameters:
        - $ref: '#/components/parameters/SchoolContext'
        - $ref: '#/components/parameters/DeviceId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Aparelho revogado, inclusive em repetição idempotente
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceState' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    adultOidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/.well-known/openid-configuration
    deviceToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-device-token
  parameters:
    SchoolContext:
      name: X-School-Id
      in: header
      required: true
      description: Contexto selecionado; a autorização é sempre reconferida no banco.
      schema: { $ref: '#/components/schemas/Id' }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
...
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
  responses:
    Problem:
      description: Erro seguro e correlacionável
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Id:
      type: string
      pattern: '^[a-z0-9][a-z0-9_-]{2,63}$'
    Timestamp:
      type: string
      format: date-time
...
    CreatePairingCodeRequest:
      type: object
      additionalProperties: false
      required: [classroomId]
      properties:
        classroomId: { $ref: '#/components/schemas/Id' }
    PairingCode:
      type: object
      additionalProperties: false
      required: [pairingId, code, expiresAt]
      properties:
        pairingId: { $ref: '#/components/schemas/Id' }
        code: { type: string, pattern: '^[23456789A-HJ-NP-Z]{4}-[23456789A-HJ-NP-Z]{4}$' }
        expiresAt: { $ref: '#/components/schemas/Timestamp' }
    RedeemPairingRequest:
      type: object
      additionalProperties: false
      required: [code, installationId, appVersion, architecture, viewportWidthDp, viewportHeightDp]
      properties:
        code: { type: string, minLength: 8, maxLength: 9 }
        installationId: { type: string, pattern: '^[A-Za-z0-9._:-]{16,128}$' }
        appVersion: { type: integer, minimum: 1 }
        architecture: { type: string, enum: [ARM64, UNIVERSAL, UNKNOWN] }
        viewportWidthDp: { type: integer, minimum: 240, maximum: 2000 }
        viewportHeightDp: { type: integer, minimum: 320, maximum: 3000 }
    DeviceCredential:
      type: object
      additionalProperties: false
      required: [deviceId, deviceToken, schoolId, classroomId, issuedAt]
      properties:
        deviceId: { $ref: '#/components/schemas/Id' }
        deviceToken: { type: string, minLength: 64, maxLength: 256, writeOnly: true }
        schoolId: { $ref: '#/components/schemas/Id' }
        classroomId: { $ref: '#/components/schemas/Id' }
        issuedAt: { $ref: '#/components/schemas/Timestamp' }
    DeviceState:
      type: object
      additionalProperties: false
      required: [deviceId, status]
      properties:
        deviceId: { $ref: '#/components/schemas/Id' }
        status: { type: string, enum: [ACTIVE, REVOKED] }
...
    Problem:
      type: object
      required: [type, title, status, code, safeMessage, correlationId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { $ref: '#/components/schemas/Id' }
        safeMessage: { type: string, maxLength: 280 }
        correlationId: { type: string, maxLength: 128 }
        issues:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              code: { $ref: '#/components/schemas/Id' }